Windows build fix

+++ cmake/Coverage.cmake
@@ -13,17 +13,19 @@ function(CleanCoverage target)
 endfunction()
 
 function(AddCoverage target)
-    find_program(LCOV_PATH lcov REQUIRED)
-    find_program(GENHTML_PATH genhtml REQUIRED)
-    add_custom_target(coverage-${target}
-            COMMAND ${LCOV_PATH} -d . --zerocounters
-            COMMAND $<TARGET_FILE:${target}>
-            COMMAND ${LCOV_PATH} -d . --capture --ignore-errors mismatch -o coverage.info
-            COMMAND ${LCOV_PATH} -r coverage.info '/usr/include/*' '*googlemock/*' '*googletest/*'
-            -o filtered.info
-            COMMAND ${GENHTML_PATH} -o coverage-${target}
-            filtered.info --legend
-            COMMAND rm -rf coverage.info filtered.info
-            WORKING_DIRECTORY ${CMAKE_BINARY_DIR}
-    )
+    find_program(LCOV_PATH lcov)
+    find_program(GENHTML_PATH genhtml)
+    if(LCOV_PATH AND GENHTML_PATH)
+        add_custom_target(coverage-${target}
+                COMMAND ${LCOV_PATH} -d . --zerocounters
+                COMMAND $<TARGET_FILE:${target}>
+                COMMAND ${LCOV_PATH} -d . --capture --ignore-errors mismatch -o coverage.info
+                COMMAND ${LCOV_PATH} -r coverage.info '/usr/include/*' '*googlemock/*' '*googletest/*'
+                -o filtered.info
+                COMMAND ${GENHTML_PATH} -o coverage-${target}
+                filtered.info --legend
+                COMMAND rm -rf coverage.info filtered.info
+                WORKING_DIRECTORY ${CMAKE_BINARY_DIR}
+        )
+    endif()
 endfunction()
